Move pasteToNewMemoryVector method to QgsClipboard
Slightly reduces code in qgisapp.cpp, and makes this method
testable without requiring QgisApp
